Add UV
This changes the `package.json` file
so that it will install UV when `yarn install` is ran.

It also configures blacklight to use the `uv` partial
on the show page.

An initializer is added to configure, via an ENV variable,
what the IIIF manifest url is for a work.

The basic `uv.html` and `uv-config.json` files are added
to the `uv` folder in `config` so that they can be customized.

Connected to emory-libraries/dlp-lux#61